feat: add reusable seeding utility
