+++ LGHackerton/utils/seed.py
@@ -1,0 +1,17 @@
+import random
+import numpy as np
+
+try:
+    import torch
+except Exception:
+    torch = None
+
+
+def set_seed(seed: int) -> None:
+    random.seed(seed)
+    np.random.seed(seed)
+    if torch is not None:
+        torch.manual_seed(seed)
+        torch.cuda.manual_seed_all(seed)
+        torch.backends.cudnn.deterministic = True
+        torch.backends.cudnn.benchmark = False
